requirements.txt
# Core ML dependencies
torch>=2.1.0
transformers>=4.36.0
peft>=0.7.0
datasets>=2.14.0
accelerate>=0.25.0
bitsandbytes>=0.41.0
pyyaml>=6.0.0

# Inference & API
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
gradio>=4.0.0

# ITSM integrations
requests>=2.31.0
python-dotenv>=1.0.0

# Monitoring & evaluation
tensorboard>=2.15.0
scikit-learn>=1.3.0

# Dev tools
pytest>=7.4.0
black>=23.12.0
flake8>=6.1.0
